refactor(dcc): tighten preprocessor and regalloc boundaries
Complete the medium-priority compiler architecture work by removing repeated loop-register-allocation scaffolding, extracting the preprocessor conditional-expression evaluator, and moving limited-consumer contracts out of the umbrella header.

Loop register allocation:
- Share candidate selection and speculative dispatch across for, while, and do-while AST wrappers.
- Add a LoopRegallocState snapshot for the frame and sequence counters restored after declined attempts.
- Share BC prime and spill emission between read-only and write-candidate trials while preserving instruction order and fallback behavior.

Preprocessor ownership:
- Move the 465-line #if/#elif precedence evaluator from dcc_preproc.c into dcc_pp_expr.c.
- Keep the evaluator cursor, recursion depth, and all 19 precedence helpers file-local.
- Expose only pp_eval_simple_expr as the evaluator entry point.
- Reduce dcc_preproc.c from roughly 3,003 to 2,546 lines.

Internal interfaces:
- Add dcc_preproc_internal.h for macro-definition and conditional-expression APIs used by the driver and preprocessor modules.
- Add dcc_regalloc_internal.h for speculative-regalloc globals, trial entry points, and verification helpers.
- Remove those limited-consumer declarations from dcc.h, reducing it from 1,259 to 1,172 lines.
- Keep foundational lexer APIs in dcc.h because they are consumed across the parser rather than owned by a narrow subsystem.
- Register the new source and headers in CMake.

Code-quality impact:
- Future loop forms share one promotion policy instead of copying candidate/try/fallback control flow.
- Rollback fields and BC setup sequences now have one implementation, reducing drift between speculative paths.
- The preprocessor evaluator has a single owner and no longer exposes its implementation ladder globally.
- Internal headers make module dependencies explicit and reduce accidental access to optimizer and macro-engine internals.

Validation:
- Each slice built cleanly and produced 0 assembly differences across 316 tests and 0 diagnostic differences across 106 diagnostic cases.
- Fast suite: 307 passed, 0 failed, 9 skipped.
- Diagnostics and performance checks passed.
- No editor diagnostics or git diff whitespace errors.
