+++ src/dcc/CMakeLists.txt
@@ -28,6 +28,7 @@ set(DCC_SOURCES
     dcc_asmname.c
     dcc_diag_emit.c
     dcc_preproc.c
+    dcc_pp_expr.c      # preprocessor #if/#elif expression evaluator
     dcc_types.c
     dcc_constexpr.c
     dcc_symbols.c
@@ -52,7 +53,8 @@ set(DCC_SOURCES
     dcc_ast_gen_stmt.c      # AST statement/switch/for emitters
 )
 
-add_executable(dcc ${DCC_SOURCES} dcc.h dcc_ast_gen_internal.h)
+add_executable(dcc ${DCC_SOURCES} dcc.h dcc_ast_gen_internal.h
+                   dcc_preproc_internal.h dcc_regalloc_internal.h)
 target_include_directories(dcc PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 
 if(CMAKE_C_COMPILER_ID MATCHES "GNU|Clang|AppleClang")
